Update to corstorm weapondef for team damage
using this as a possible solution to fix the team damage that is caused by attacks from this unit to other units that is allied with. this may provide a way for each unit that has this to better save HP and perform better when attacking in a group. 
_______________
a side effect of this will be unit longevity and persistence on the battlefield, enabling the unit to get more bang for its buck. 
part of this side effect of this is the unit will be able to shoot more often in cases as it will live longer to fire more rounds.

if this change works I suggest that we migrate this change to all units that may have this issue. 
this will not effect when the unit fires but when it fires the projectile will not team damage. this will allow more units with AOE attacks not to blow up friendly units when grouped. (IE: leveler,  fatboy, goliath, to name a few)
if this is still the case of being intended to receive team fire from these units as an explosion then thats cool too but it feels bad when shooting your own units in a group.